Changed Debussy song title
The Debussy song "Le Petit Negre" (french for "the little black boy")
is, for me, uncomfortably similar to an english slur. The piece seems to be 
published as "Le Petit Noir" nowadays, so I'm changing it to that.

The song itself is kinda minstrel-y, and while I like the style (reminds me
a lot of Joplin), all of the racist power dynamics of Debussy using this style,
and giving the song
that name aren't great. It's a minor part of the code I'm trying to show here,
and I may just get rid of this song data before publishing fully as to not have that
sort of distraction, but leaving this note
for myself and any future code-archeologists to know that I now believe it was
a ignorant move on my part to include this song so heavily in my project. ;)